requirements.txt
praw==7.7.1
pandas==2.0.3
numpy==1.24.3
torch==2.0.1
transformers==4.35.2
datasets==2.14.5
accelerate==0.24.1
huggingface-hub==0.17.3
tqdm==4.66.1
python-dotenv==1.0.0
concurrent-futures==3.1.1
